pyproject.toml
[project]
name = "genai-trials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colpali-engine>=0.3.8",
    "ipython>=8.31.0",
    "janus>=2.0.0",
    "langchain>=0.3.17",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.33",
    "langchain-deepseek-official>=0.1.0",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pypdf>=5.2.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "streamlit>=1.41.1",
    "tavily-python>=0.5.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.47.1",
    "typing-extensions>=4.12.2",
]
